Change Offsetter class to OffsetCalculator, and update corresponding files
