+++ test/enigma_test.rb
@@ -1,7 +1,7 @@
 require './test/test_helper'
+require './lib/offset_calculator'
 require './lib/shift_calculator'
 require './lib/caesar_cipher'
-require './lib/offsetter'
 require './lib/keyable'
 require './lib/datable'
 require './lib/enigma'
