+++ test/test_helper.rb
@@ -5,9 +5,9 @@ SimpleCov.start
 require 'minitest/autorun'
 require 'minitest/pride'
 
+require './lib/offset_calculator'
 require './lib/shift_calculator'
 require './lib/caesar_cipher'
-require './lib/offsetter'
 require './lib/keyable'
 require './lib/datable'
 require './lib/enigma'
